Add the "single" MIG strategy
With strategy, a give node will only expose a *single* type of MIG
device across all of its GPUs. It is designed for large deployments
(i.e. those with many GPU capable nodes into the cluster) that wish to
dedicate an entire node to a given MIG device type and continue to use
the traditional "nvidia.com/gpu" resource type to get access to them.

Signed-off-by: Kevin Klues <kklues@nvidia.com>